auth-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
email-validator==2.1.0.post1